Output templating
As the graphs we are dealing with here can become large and unwieldy,
there are now a set of jinja-templates that format the output.

These templates are currently capable of dealing with groups, subgroups,
and fields. This hierarchy is somewhat a design decision in the actual
parsing-code, rather than template-dependant.

If it turns out this design is wrong, it will have to be re-worked in
both the parser/builder, and the templates.